Problem: duplicate code between tx-validation and tx-query enclaves
Solution: extracted out the common unsealing code into a library
shared between tx-validation and tx-query
